requirements.txt
streamlit==1.31.0
tensorflow-cpu==2.12.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
pillow==9.5.0
protobuf==3.20.3
altair==5.0.1
h5py==3.8.0
